Did some code tree scrubbing.
Changed xxxGetAttrValue(), xxxSetAttrValue(), xxxAddAttr(), and
xxxExecuteMethod() to use pObjData as the type for val (or param in
the case of xxxExecuteMethod) instead of void* for the audio, BerkeleyDB,
GZip, HTTP, MBox, MIME, and Shell drivers, and found/fixed a 2-byte buffer
overflow in objdrv_shell.c (line 1046).

Also, the Berkeley API changed in v4 in a few spots, so objdrv_berk.c is
broken as of right now.

It should be noted that I haven't actually built the audio or Berkeley
drivers, so I *could* have messed up, but they look ok. The others
compiled, and passed a cursory testing.