Added a new POP3 driver. The driver is responsible for three types of objects: the POP server object, the POP maildrop object, and the POP message object. Servers contain maildrops, which contain messages.
The driver works, but some functionality is missing. The only method
of enumerating maildrops that has been implemented is 'currlogin'. The
only method of authenticating is also 'currlogin'. In other words, the
driver will assume that the username/password to use on the POP server
are the username/password you used to log in to Centrallix. Other methods
are stubbed out.

The driver *should* handle concurrency properly, sharing one connection
between multiple open objects on the same POP server. I haven't tested
this strenuously yet.

Also, when a message object is opened, it is pulled off the
server and stored in memory in its entirety, until the object is closed.
This isn't such a good idea for e-mails with a few 15meg attatchments,
and should be changed whenever the object system is retrofitted with a
universal caching mechanism.