OSML: Adding auto type detection for structure files.
If an object has no apparent type (application/octet-stream) as determined
from the object name, and the ->Prev driver didn't indicate a type more
specific than application/octet-stream, then we probe the object to see
if it is a structure file.  If so, we snarf the type from the structure
file's top-level group and go with that, as long as it is valid and is a
valid subtype of "system/structure".

Caveat:  This only works for structure files.  We may add magic number
detection logic later on that will help us with other file types too.

Note:  I removed the type-detection logic for "_DB" endings from types.cfg.
This is because the objects that was set up for, sybase and mysql, will
be better served by this auto-detection logic anyhow.