Firefox 4.0 Updates - Fix for FF4 bug that broke dynamic components.
Firefox 4.0 introduces async scripts.  If you tag the script as
async = false, it is supposed to disable this behavior and guarantee
that scripts are run in their insertion order.  However, this is
apparently *NOT* happening, at least when the scripts are a mix
of src= and text scripts.  This broke the ability of Centrallix
to know when all of the scripts have loaded, before invoking the
startup function for a dynamically loaded component, because FF4
executed the last script inserted before it had executed earlier
scripts.

Firefox 4.0 provides a way for a script completion event (the
onbeforescriptexecute event), but this event is not cross-browser
and is only available on FF4 (Gecko 2).

Our solution is to maintain an array of loaded scripts, which is
updated by a small line at the end of each .js file.  This works
well both in FF4 and in FF3.x.