Bugfix: previous fix for local file open mode caused regression
This issue needs to be reviewed; this fix is a quick one, but
techincally can allow the file to be open read/write when it
should be readonly.

The problem the previous fix caused was that a CSV row could not
be updated/inserted/deleted because obj->Prev was readonly even
though the CSV row was read/write.

This keeps the previous issue fixed, though, because the previous
fix had to do with O_TRUNC getting passed to the file open when
it should not have been.  O_ACCMODE masks that out.