Added HTTP stating of content charset type.
Also added a new utility module with the prefix chr.
The header is utility/charsets.h and it contains a
function for looking up the charset to be used by a
module and some macros to specify attribute names
that are allowed in charsetmap.cfg.  Added one other
macro to define the key in centrallix.conf that
specifies the charsetmap file.