Adding Crypto-JS v2.5.1 to centrallix-os
Crypto-js provides encryption and hash function support for javascript.
It is licensed under a 3-clause BSD license, which is compatible with
our LGPL 2.1.  This is being imported into our git, so we will have to
keep this up to date as important releases are made.  We are not remote
linking to an online copy since that would 1) have security issues,
2) prevent the use of Centrallix in a non-internet-connected network,
and 3) cause us trouble should the maintainers change an API.

See: http://code.google.com/p/crypto-js/