Fix for MTASK scheduler on 64bit systems
This is a temporary fix that resolves a problem in the MTASK scheduler
where sizeof(long) == 8.  Timer variables are reverted to 32bit in this
patch, returning to the behavior of the scheduler on 32bit systems.
The scheduler's timing variables need to be overhauled further, but this
patch fixes the issue for now.