Change centrallix.init.in so init.d/centrallix has pid file info
These changes were made for centrallix to work as a service with
systemd.  For systemd to work, it needs to know the pid file, and
centrallix needs to make a pid file.

The centrallix.init.in file is what the init.d/centrallix startup
script is generated from.  Added a -p /var/run/centrallix.pid line
to the centrallix server startup.  Also changed the comment that
says where the pid file is located.