Changes to the build system for Mac OS X compatibility
This removes the flag -Wno-unused-but-set-variable, which was used
when compiling json-c since it is not recognized by Clang.

The behavior for undefined symbols while compiling shared libraries
has been set to dynamic_lookup.

The call determine the version of the program in configure.ac now
calls tr, since the Mac OS X version of echo is missing the -n
option to trim the trailing newline character.