First draft of .c files for chart widget
The wgtdrv file is done as far as I know.
In that file I was mainly adding virtual subwidgets for the axis and series.

The htmlgen file is a complete rewrite.
The code has been refactored to be very functional and clean.
It's successfully passing some basic params through to the javascript and it should be easy to add more.