Support for pie, bar, line, and scatter charts
This involved some small tweaks to the htmlgen file and a complete rewrite of the js file.
The javascript file uses functional techniques to manipulate config parameters and chart data.
As long as it's given adequate data, the javascript is able to create a chart even if it's not told how to interpret the data.