Added several configuration params and a few tweaks
We were able to add a lot of configuration entries and documented them all.
There are also a few small tweaks in here to improve the polish of the widget.

This should be the final commit for code-a-thon.
The widget is not fully fleshed-out, but the structure and basic parameters are there.