Move conversion to old money representation for JSON into separate function to enable test.
However, even though most of the business logic is being tested now, nht_i_RestWriteAttrValue (or public callers) still is not tested at all. It would be trivially easy to test with mocks/stubs, but is quite complex to test otherwise, and so I judged that it wasn't worth it to spend a lot of time writing a complete test right now when we might be implementing a test framework with mocks/stubs relatively soon. Please, future me, save us from the screaming void.